feat: refactored some code
