+++ assets/react/v3/entries/course-builder/services/quiz.ts
@@ -18,20 +18,36 @@ export interface QuizQuestionOption {
   matchedTitle?: string;
   fillinTheBlanksCorrectAnswer?: string[];
 }
-export interface QuizQuestion {
+
+// Define a base interface for common properties
+interface BaseQuizQuestion {
   ID: string;
   title: string;
   description: string;
-  type: QuizQuestionType;
   answer_required: boolean;
   randomize_question: boolean;
   question_mark: number;
   show_question_mark: boolean;
-  muliple_correct_answer: boolean;
-  image_matching: boolean;
   markAsCorrect?: string | string[];
   options?: QuizQuestionOption[];
+  answerExplanation: string;
 }
+
+interface MultipleChoiceQuizQuestion extends BaseQuizQuestion {
+  type: 'multiple-choice';
+  muliple_correct_answer: boolean;
+}
+
+interface MatchingQuizQuestion extends BaseQuizQuestion {
+  type: 'matching';
+  image_matching: boolean;
+}
+
+interface OtherQuizQuestion extends BaseQuizQuestion {
+  type: Exclude<QuizQuestionType, 'multiple-choice' | 'matching'>;
+}
+
+export type QuizQuestion = MultipleChoiceQuizQuestion | MatchingQuizQuestion | OtherQuizQuestion;
 
 const mockQuizQuestions: QuizQuestion[] = [
   {
@@ -40,16 +56,15 @@ const mockQuizQuestions: QuizQuestion[] = [
     title: 'Trust me I am True / False',
     description: 'This is a true false question',
     options: [
-      { ID: '1', title: 'True'},
+      { ID: '1', title: 'True' },
       { ID: '2', title: 'False' },
     ],
     answer_required: false,
     question_mark: 1,
     randomize_question: false,
     show_question_mark: true,
-    muliple_correct_answer: false,
-    image_matching: false,
     markAsCorrect: '1',
+    answerExplanation: '',
   },
   {
     ID: '3',
@@ -59,7 +74,7 @@ const mockQuizQuestions: QuizQuestion[] = [
     options: [
       { ID: '1', title: 'Option 1' },
       { ID: '2', title: 'Option 2' },
-      { ID: '3', title: 'Option 3',},
+      { ID: '3', title: 'Option 3' },
       { ID: '4', title: 'Option 4' },
     ],
     answer_required: false,
@@ -67,8 +82,8 @@ const mockQuizQuestions: QuizQuestion[] = [
     randomize_question: false,
     show_question_mark: false,
     muliple_correct_answer: false,
-    image_matching: false,
     markAsCorrect: '3',
+    answerExplanation: 'This is the answer explanation',
   },
   {
     ID: '4',
@@ -79,8 +94,7 @@ const mockQuizQuestions: QuizQuestion[] = [
     question_mark: 1,
     randomize_question: false,
     show_question_mark: true,
-    muliple_correct_answer: false,
-    image_matching: false,
+    answerExplanation: 'This is the answer explanation',
   },
   {
     ID: '5',
@@ -91,11 +105,8 @@ const mockQuizQuestions: QuizQuestion[] = [
     question_mark: 1,
     randomize_question: false,
     show_question_mark: true,
-    muliple_correct_answer: false,
-    image_matching: false,
-    options:[
-      { ID: '1', title: 'Fill the {dash} in time', fillinTheBlanksCorrectAnswer: ['gap'] },
-    ]
+    options: [{ ID: '1', title: 'Fill the {dash} in time', fillinTheBlanksCorrectAnswer: ['gap'] }],
+    answerExplanation: 'This is the answer explanation',
   },
   {
     ID: '6',
@@ -106,8 +117,7 @@ const mockQuizQuestions: QuizQuestion[] = [
     question_mark: 1,
     randomize_question: false,
     show_question_mark: true,
-    muliple_correct_answer: false,
-    image_matching: false,
+    answerExplanation: 'This is the answer explanation',
   },
   {
     ID: '7',
@@ -118,7 +128,6 @@ const mockQuizQuestions: QuizQuestion[] = [
     question_mark: 1,
     randomize_question: false,
     show_question_mark: true,
-    muliple_correct_answer: false,
     image_matching: false,
     options: [
       { ID: '1', title: 'Option 1', matchedTitle: 'Matched Option 1' },
@@ -126,7 +135,8 @@ const mockQuizQuestions: QuizQuestion[] = [
       { ID: '3', title: 'Option 3', matchedTitle: 'Matched Option 3' },
       { ID: '4', title: 'Option 4', matchedTitle: 'Matched Option 4' },
     ],
-    markAsCorrect: '1'
+    markAsCorrect: '1',
+    answerExplanation: 'This is the answer explanation',
   },
   {
     ID: '9',
@@ -137,15 +147,14 @@ const mockQuizQuestions: QuizQuestion[] = [
     question_mark: 1,
     randomize_question: false,
     show_question_mark: true,
-    muliple_correct_answer: false,
-    image_matching: false,
     options: [
       { ID: '1', title: 'Option 1' },
       { ID: '2', title: 'Option 2' },
-      { ID: '3', title: 'Option 3',},
+      { ID: '3', title: 'Option 3' },
       { ID: '4', title: 'Option 4' },
     ],
-    markAsCorrect: '3'
+    markAsCorrect: '3',
+    answerExplanation: 'This is the answer explanation',
   },
   {
     ID: '10',
@@ -153,18 +162,17 @@ const mockQuizQuestions: QuizQuestion[] = [
     title: 'Order is not chaos!',
     description: 'This is an ordering question',
     options: [
-      { ID: '1', title: 'Option 5', },
+      { ID: '1', title: 'Option 5' },
       { ID: '2', title: 'Option 6' },
-      { ID: '3', title: 'Option 7'},
+      { ID: '3', title: 'Option 7' },
       { ID: '4', title: 'Option 8' },
     ],
     answer_required: false,
     question_mark: 1,
     randomize_question: false,
     show_question_mark: true,
-    muliple_correct_answer: false,
-    image_matching: false,
-    markAsCorrect: '3'
+    markAsCorrect: '3',
+    answerExplanation: 'This is the answer explanation',
   },
 ];
